Fix scary type issues and add online tests.
Found a few places (thanks compiler warnings!) where an FSyncTree*
was being passed to an id<Node> method. Added online tests as requested
but I'm not sure if it will screw up the expected event count.

I can't run tests manually since the serverside changes aren't public
and I don't know how to make the emulator work locally, so we'll push,
pray, and trust Travis. If this doesn't work I may not be able to
finish the PR.